MATLAB Figure Arrangement Script
The purpose of this script is to arrange MATLAB figures in a neat grid on the screen, rather than allowing the figures to appear in overlapping windows. This allows the easier comparison of multiple figures. The code is commented to allow easy navigation. An example is shown in the pdf.