Fast Decoupled Flow Method Script 1
Read the pdf explaining the purpose of the code. 

This MATLAB script is designed to apply the Fast Decoupled Flow Method (FDFM) to simple systems with a single given voltage and phase angle at bus 1 (slack bus) connected via a single line of known impedance to bus 2 of known apparent power demand. The code will likely be expanded in future, so this is just the first version.